requirements.txt
asgiref==3.4.1
Babel==2.9.1
captcha==0.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2
django-allauth==0.47.0
django-colorfield==0.6.3
django-etc==1.3.1
django-filter==21.1
django-hitcount==1.3.5
django-money==2.1.1
django-oauth2==3.2
django-phone-field==1.8.1
django-resized==0.3.11
django-social-auth==0.7.28
django-social-share==2.2.1
django-taggit==2.0.0
djangorestframework==3.13.1
httplib2==0.20.2
idna==3.3
oauth2==1.9.0.post1
oauthlib==3.1.1
Pillow==9.0.0
py-moneyed==1.2
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-openid==2.2.5
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
shortuuid==1.0.8
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
urllib3==1.26.8
